moving get_fmap_type to utils
